Templates Module
After much deliberation, I've decided the most intutive and most extensible way to handle templates was to create a module (templates.py) which creates a dictionary for every template type which contatins the name of the template and its associated html generating function. The compile function tells templates.py to create the dictionaries and then imports the layouts dictionary. Developers are then free to import template dictionaries as need be.