New Templating System
This commit redid the templating system so that developers can import
templates directly form the main executable which will make it easier to
package the program into a single executable later. This commit also
changes the way blocks are imported from a block dict, to a block
function that accepts the block name and block data. The README and
examples have been updated to explain how to use this new system